IMPALA-1264: Re-enable test_session_experation.py
This has only failed once as far as I can tell, and we don't have any log
files for the failure.  Nor have I been able to repro.  So, let's increase the
test timeout to get a bit more slop in the case of the occasional long sleep
on the impalad side.  If there is a bug where a session is stuck in the
checked-out state, the test will still find it.

Change-Id: Ib63a8a83d9c3e772787dc548473868c58f950eb8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4494
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins