IMPALA-1100: AnalysisException when CTAS has a WITH clause and inline view
This commit fixes the issue where a CTAS statement with a WITH clause and
an inline view throws an AnalysisException. The problem was that during
the analysis of the WITH clause, the globalState of the analyzer used in
the CTAS was poluted because the same conjuncts were registered multiple
times.

Change-Id: I4f27aeb0f2107d565c80ac9e3eef7d0387865bdc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4574
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins