IMPALA-1354: Disallow constant expressions in order by and partition by exprs
Postgres and Oracle allow constant expressions in the order by and partition exprs
which we don't currently handle properly when there are no other slots referenced.
In this case, the FE still produces a SortNode (though there's nothing to sort)
with an empty sort tuple and the BE crashes. Given that this is a strange corner
case that isn't really useful, we're just disallowing constant exprs in order by
and partition by exprs for now.

Change-Id: Ib2bcf8896a1fb7100c19e3a2bf4bfb92514feab9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4765
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: jenkins