IMPALA-1451: Empty Row in HBase triggers NPE in Planner
If for some reason an empty row is returned during a scan of an HBase
table, we have to make sure it is ignored.  Otherwise, it will trigger a
NPE when accessing either the row key or the value.

Change-Id: Ieab4b765a4129236d348b711b36fd666979811dc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5145
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: jenkins
(cherry picked from commit 4009b927a961c744c4e6435820593c57190226ba)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5171