IMPALA-1617: Bug fix to avoid crash in PHJ::BuildHashTables() on low memory
The partition spilling code of PHJ had a wrong DCHECK that would fail in cases of very low
memory, where we would have memory to allocate buffers but not the hash table. In
particular, the spill code would DCHECK if the partition was still using small buffers,
which is wrong as partitions with small buffers can still spill.

This patch also adds a test that makes sure that TPC-H Q4 (that contains a join and a
group by) either executes correctly or gracefully fails in a set of relatively small mem
limits. Currently when compute stats have been run to both lineitem and orders tables,
TPC-H Q4 needs a minimum of 150MB to run successfully.

Change-Id: I997718384832bd6bf469cd608e9f692b6d727c4f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5752
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: jenkins