IMPALA-1630: Fix race with scanner thread leading to failed query returning OK
HdfsScanNode::GetNextInternal() needs to check the shared status_ after
returning from the materialized_row_batches_ blocking queue.  Otherwise,
the scanner thread could have set status_ to an error after status_ was
checked in GetNextInternal().

The early status_ check is unnecessary because the scanner thread will
always call SetDone() and shutdown the blocking queue after setting the
shared status_, and so the new status_ check is sufficient. This means
we no longer need to take this lock on every GetNextInternal call.

Change-Id: I6dbbd5f28d754877896d4006152bcf86925f2765
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5662
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins