Re-enable TPCDS planner tests.
These tests had been 'temporarily' disabled when moving the TPCDS schema
on CDH5 to a partitioned store_sales with DECIMAL. The intention was to
to re-enable these tests shortly after the schema change, but it was never
actually done.

Process to restore this test:
I started off with the tpcds-all.test file from our CDH4 branch and ran
it on CDH5. I investigated the following plan differences:
- Table sizes are slightly different on CDH5
- Several plan changes, e.g., join order, analytic order. All plan differences
  are due to the size difference between DECIMAL and FLOAT. The CDH4 tables
  use FLOAT, and the CDH5 tables use DECIMAL. Some plans had aggregates on those
  columns, and the size difference between, e.g., a DECIMAL(38,2) and a DOUBLE
  was significant enough to change the plan choice in several instances.

I concluded that all the plan differences are legitimate and should be accepted.

Change-Id: I11f36a543e9a5041d569c6f633fdfd296b72d31e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5672
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins