IMPALA-377: Log Rotation for Impala Processes
This patch enables a simple form of log rotation for Impala. Using the
new command line flag:

    -max_log_files=XX

the specified number of log files is kept for each severity level and
all older log files are removed. The number has to be greater than 1 to
keep at least the current log file open. If set to 0, all log files will
be retained and log rotation is effectively disabled. The rotation is
checked with every interval of "logbufsecs" that is set to a default
value of 5s.

Testing is limited as logging can only be initialized once per process.

Change-Id: Iafb0f165bf705d2aa0d6a7f2f440662329af6feb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5706
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: jenkins