IMPALA-1674: Fix serious memory leak in TSaslTransport
TSaslTransport uses a memory buffer (TMemoryBuffer) object to stage
bytes that the layer above it will only partially read during the
current operation. However, that buffer is never shrunk when the bytes
are eventually read in a further read() call, and the result is that the
buffer expands to contain *all* the bytes that were staged in it during
the lifetime of the transport, which is the same as the connection
lifetime.

So, in secure clusters, backend<->backend connections would leak memory
- sometimes slowly, sometimes very quickly - as the buffer became
extremely large. The leak was not apparent on LDAP secured clusters
because only the client<->Impala connection was using TSaslTransport,
and that connection was a) often short-lived and b) sent very little
data *to* the Impala daemon. The leak only occurs on the C++'s
implementation's read() path (which calls write() to the memory buffer).

This patch addresses the problem by checking the size of the 'junk'
section in the buffer (see below). This corresponds to the number of
already-read bytes. If that goes above a certain size, we consider
garbage collecting the buffer and replacing it with a new one.

+----------+---------------------+----------+
|   Junk   |      To be read     |   Free   |
+----------+---------------------+----------+

It's only safe to do this garbage collection when the 'to be read'
section is empty, otherwise resizing, which is a destructive operation,
would lose 'live' bytes. Alternatively, we could copy out the pending
section into the new buffer, but this would complicate things only for
the purpose of keeping the buffer size down *during* an RPC, since at
the end of the RPC the 'to be read' section will be 0-sized and the
buffer will be garbage collected if needed.

Change-Id: Ic5eecdd704fb09734431f682af0919581861b12a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5789
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: jenkins