S3: Simplify loading of file block metadata
Prior to this change, file block metadata is loaded in three stages:

1) Each new file gets a new FileDescriptor object.
2) After all new FileDescriptors are created, we loop through them all
   and get the block information for each.
3) Then, once all the blocks are collected, we call to DFS to get the
   disk volume IDs.

This change combines stages 1 & 2 into a single step.  This is simpler.
Also, this allows us to get the FileStatus only once per file, rather
than twice.  Finally, this sets up an S3 catalog change to compute file
blocks for S3 (file blocks need to be determined differently than how
they are on DFS), and some information is needed that is available only
at step 1 but not step 2.

More cleanup is possible in this code, but this is a good step in the
right direction.

Change-Id: I9a0fa2e5725fa64b180499e790a3afca47ae51c1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5855
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins