IMPALA-1708: Don't warn about stats with 0 rows but > 0 files
A small bug in HdfsScanNode caused a partition that had 0 rows to be counted
as missing stats. This showed up only when that partition had some files
in it, because otherwise there'd be no scan range for that partition and
it would be ignored. This patch also fixes a bug where the number of
partitions without stats wasn't zeroed at the beginning of
HdfsScanNode.computeStats(), so the total was double counted for some
queries which called computeStats() twice.

A new test case constructs a table with a partition with 0 rows but
containing an empty file, and confirms that the warning message is not seen.

Change-Id: I7a93fd929175b8c616d2ff42983c5969c66e868f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5864
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins