S3: Fix loading avro schema from S3
The avro schema loading code was accepting only the hdfs filesystem
scheme, as that was hardcoded.  Though even then the checks weren't
really sufficient for hdfs (the URI could have pointed to a
different/invalid namenode).  Fix this code up to work so that the avro
schemas can be loaded from any accessable Hadoop filesystem.

Change-Id: Iaf2d82bdc686652b83db06281fef5a1a2363bcb8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5874
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins