S3: Refactor PlannerTest so we can have S3-specific tests
Many of the PlannerTest test cases won't run on S3 because they depend on:

a) Insert, which isn't currently supported on S3, or
b) HBase, which we aren't setting up when testing on S3, or
c) HDFS specific block characteristics, when checking scan ranges.

So, let's not bother running the normal planner tests on S3.  However,
in order to add an S3A scan range test, I'd still like to be able to use
the PlannerTest infrastructure code.  So, split the code up.

In this change I've added just a stub S3 test case so we can see what
it'll look like.  I'll add the real test case in a following change
where it'll be changing how scan ranges work on S3.

Change-Id: I58ed4ee0f4f7c173741ce50a7272258037787fb4
Reviewed-on: http://gerrit.cloudera.org:8080/44
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins