IMPALA-1712: Remove "stale metadata" warning
The "Block locality metadata ... may be stale" message in Impala 2.1 can
occur apparently spuriously. Running INVALIDATE METADATA as suggested
does not fix the problem. The result is understandable frustration.

This patch removes that message entirely until we understand the root
cause. Until we do, there's no point writing messages even to the log:
this is not actionable information for operators or users, and may be
incorrect anyhow. We can still detect this condition by checking the
BytesReadRemoteUnexpected counter in the query profile.

Change-Id: I3b923f9adaccc055c3b266cd60cf091266c2acb6
Reviewed-on: http://gerrit.cloudera.org:8080/58
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins