Restructure query detail webpages
This patch adds the /query_plan endpoint, which uses the Dagre and D3 JS
libraries to render the plan tree. The tree auto-refreshes once every
second, and the edges are labelled with the output cardinalities of each
operator. Plan fragment boundaries are marked by a different edge
colour, and nodes that belong to the same fragment are all coloured the
same.

This patch also restructures the summary and profile pages under one new
'details' heading that gives the user a selection of tabs to choose
between, each with some relevant query information. Those tabs contain
the plan tree, the query statement, query summary, plan text and the
profile.

Change-Id: I759970bf7a41874435baff700025ee0a3407d222
Reviewed-on: http://gerrit.cloudera.org:8080/68
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins