Fix and restore least/greatest expr test
I accidently commented out a test as part of d448f5 while doing final
cleanup after a +2. The test was failing because it relies on the
ordering of an unordered map. Just changing to an ordered map fixes the
test. For whatever reason an unordered map worked on Ubuntu 12.04 but
fails on 14.04.

Change-Id: Ia1bf04c53ee5cbcfc2d3f11e222ce4d47f1973c1
Reviewed-on: http://gerrit.cloudera.org:8080/69
Reviewed-by: Casey Ching <casey@cloudera.com>
Tested-by: Internal Jenkins