CDH-24093: Impala should produce column lineage info needed by Navigator
This change adds support for column lineage logging in Impala to be
consumed by Navigator. This feature is disabled by default and is enabled
by setting the -lineage_event_log_dir flag. When lineage logging is
enabled, the serialized column lineage graph is computed for each query
and stored in a specialized log file in JSON format.

Change-Id: Ib8d69cdbcc435be1e9c9694998c1d33ec1245b10
Reviewed-on: http://gerrit.cloudera.org:8080/70
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>
Tested-by: Internal Jenkins