IMPALA-1794: Fix infinite loop opening/closing file w/ invalid metadata
Files which are smaller than the cached size (e.g. file truncated without
running invalidate metadata) cause the ScannerContext to enter an infinite
loop where the file is continuously opened and closed as we attempt to read
more expected bytes.

This fixes the issue by identifying files shorter than expected and updating
the expected file length and using that to determine the eof and eos.

Change-Id: Ibde588eb0c73a94f2bb9022b3fe42be97681dcb2
Reviewed-on: http://gerrit.cloudera.org:8080/71
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins