IMPALA-1646: Don't compile with -msse4.2
Currently, Impala binaries are compiled with -msse4.2 so that the SSE
4.2 intrinsics can be used on the SSE-optimized paths.  While these
paths correctly have runtime checks for SSE 4.2 support, the -msse4.2
flag allows the compiler to automatically emit SSE 4.2 instructions.
Compilers have become smart enough to do so, and so various builds of
Impala don't work on CPUs that lack SSE 4.2 (or 4.1, etc) support.

Fix this by defining our own implementations of the intrinsics so that
we don't have to use -msse4.2 during the native compile.  This allows us
to continue to inline these paths.  We were already cross-compiling for
SSE 4.2 and non-SSE 4.2, so we continue to leverage that for IR.

Additional testing:
- spot check the SSE 4.2 optimized paths and see that the new code is
  the same as old code.
- mask support for SSE 4.2 and run tests to verify non-SSE 4.2 code
  paths work.  (A follow on change will add an option to do this so
  we can add it to our regression tests).
- Spot check a few locations where we know the compiler was
  auto-emitting SSE 4 instructions and verify that's not happening now.
  (Unfortunately, we don't have any such machines in our test cluster).

Change-Id: I85afd8f0c880e4dfe86b110e09061593f7698619
Reviewed-on: http://gerrit.cloudera.org:8080/111
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins