IMPALA-1841, IMPALA-1873: Making the block mgr thread safer
This patch fixes some thread-safety problems in the block mgr that would probably appear
as DCHECK(Validate()) errors, as those reported in IMPALA-1841 and IMPALA-1873.
In particular:
- Converts the is_cancelled_ flag back to bool and protects it with the lock_ usually as
part of larger critical sections.
- Makes sure that PinBlock() is modifying the shared total_pinned_buffers_ and
free_io_buffers_ and unpinned_blocks_ lists as well as is calling WriteUnpinnedBlocks()
with protection.
- Moves a called to block->Validate() that needs protection with the block mgr lock_
outside the GetUnusedBlock() call that it is supposed to not need locks.

Will add unit tests for the block mgr with multiple threads in a separate patch.

Change-Id: I8cb9e0d1931e6f67c728252ff0c2b8bd4bc2bcc9
Reviewed-on: http://gerrit.cloudera.org:8080/237
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins