Checking return status of WriteUnpinnedBlocks()
There were two callers of BufferedBlockMgr::WriteUnpinnedBlocks() that were not checking
its return status. This patch fixes a bug where one of the two callers does not wake up
potential waiting threads when WriteUnpinnedBlocks() returns errors. This could result in
queries hanging. We now check the returned status and if it is not ok() we wake up any
waiting threads and fail the query.

Change-Id: I1b1f837b5fd8323cb9aa5d46aef073e3858ada81
Reviewed-on: http://gerrit.cloudera.org:8080/213
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins