Defensively ensure that tables cannot have the same name in DDL stress test
test_ddl_stress.py uses uuid4() to create a unique identifier for each
table. Since these tables are generated concurrently, we should be alert
to the possibility that concurrent calls to uuid4() give the same
result, leading to test failures.

It's not clear if this case is possible on Linux, but anecdotal reports
suggest that certain kernel versions (and some versions of OS X) do not
guarantee low collision probability in the face of concurrent readers.

Change-Id: I6af5d2ca8c1f46ead3ec2c9216b2ab9c2291bd83
Reviewed-on: http://gerrit.cloudera.org:8080/256
Reviewed-by: Ishaan Joshi <ishaan@cloudera.com>
Tested-by: Internal Jenkins