Improve partitioned agg hash table cleanup
This improves the partitioned agg cleanup of tuples in hash
tables. In cleaning up the hash tables, any intermediate
tuples need to have Serialize/Finalize called in order to
release any resources that may have been allocated by UDAs.

The code currently allocates new output tuples for every
intermediate tuple that needs to be cleaned up. This is both
unnecessary and potentially dangerous because the memory is
allocated from mem_pool_ and, during the course of this
cleanup, mem_pool_ is occasionally freed to avoid
accumulating memory.

This was actually a bug in the old aggregation node (fixed
in commit 778d8caafc66dc20a90e096f0ab686e5433fd94f on
cdh5-2.2.0_5.4.x) because mem_pool_ was used during the hash
table build to create all intermediate tuples, thus clearing
the mem_pool_ would free memory that we again try to clean
up in Close(). While this is not currently happening in the
partitioned agg, we should avoid all of the allocations from
mem_pool_ so we don't have to worry about accumulating
memory and possibly introducing bugs in the future if we
start using mem_pool_ in other places.

Change-Id: I50300401e54cdde1a5b77b9aab65e3dcd3704a22
Reviewed-on: http://gerrit.cloudera.org:8080/289
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins