Add flag to enable building with ccache
Adds support in the be CMakeLists.txt to enable building
with ccache. This is enabled by default if ccache is installed,
but can be disabled by setting the DISABLE_CCACHE environment
variable.

This is primarily helpful to avoid having to modify the PATH
to execute ccache (e.g. symlinks from gcc,clang,etc. to ccache)
and to hide some flags that have to be set to work with clang.

Change-Id: If9af5312254a43ca7dd85d3ac022a849be97738c
Reviewed-on: http://gerrit.cloudera.org:8080/290
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: Internal Jenkins