Add a timer for hash table resizes in PAGG
This patch adds a timer for the hash table resizes in PAGG. While
investigating the perf of the new hash table we need to know how much
time we are spending in resizes. For example, this timer will help us
evaluate how important it is to have more accurate estimations of the
initial hash table sizes in PAGG (now we init all PAGG hash tables
with 1024 buckets). It will also give us an estimate on how fast the
resize function is, and how critical its perf is.

Change-Id: I49af0c9568a54bec0720f48074bf04af517b903a
Reviewed-on: http://gerrit.cloudera.org:8080/306
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins