Add timestamp to TScanRangeParams to perform SNAPSHOT scans in Kudu
Scans from impala should always be snapshot scans. However before the
scan starts a timestamp should be selected so that all scan nodes use
the same timestamp, thus guaranteeing that results returned from Kudu
are consistent.

This adds a 'timestamp' field ro the TScanRangeParams so that we can
pass that information from the frontend to the backend.

Change-Id: I43461671a14395cd7faf888d88ddfed84f86cf23
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6621
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>