gutil: Update to Kudu's latest version
On updating the client to Kudu's latest version we started hitting conflicts
on the gutil files Kudu's client exports versus the ones that Impala has in
thirdparty.

This updates the files that are used in gutil to their latest version in Kudu.
Because several files in gutil now use the PRIi64 macro, we now need the
-D__STDC_FORMAT_MACROS definition set on the build. This also sets the definition
on the core code and on the IR compiler. Impala-lzo will also require this
be added in order to compile since it uses gutil.

Change-Id: Ic618507745428e3c89288c63b4a8bda6c5e1eae7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6613
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>