Add a way to create a KuduSchema from the expression vector and table desc
On scans we can build the KuduSchema from the TupleDescriptor since the slots
refer to the path of the Kudu columns they correspond to. On
inserts/updates/deletes however this information is not available on the slots
since they might originate somewhere else. In this case we need to infer
the Kudu schema from the expressions and from the table descriptor.

This patch adds a way to build a KuduSchema extracting the types from a
the expressions vector and the column names from a KuduTableDescriptor.

Change-Id: I6e7e6c06a659b3c143894cc2463f337dcd3b73b6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6876
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>