Frontend part of Kudu inserts
This adds the frontend part of Kudu inserts, namely the KuduTableSink
and the Kudu specific processing in InsertStmt. Kudu inserts do not
support PARTITION or OVERWRITE clauses.

This patch also includes a planner test that tests that the plans for
INSERT INTO ... VALUES and INSERT INTO ... SELECT FROM ... are correctly
generated.

End-to-end tests require unrelated boilerplate code and will be included
in a follow up patch.

Change-Id: I89388622a1cfc10ced16fd944999a69986b6c6d0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6879
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>