Avoid using 'liketbl' in tests
This table's test data contains rows whose the first column
is null. As that column is select as primary key in Kudu this
makes it so we try to insert a row with a NULL key, when loading
test data, which makes data loading fail.

This patch just removes this table from the set of test tables
used in Kudu and makes it so KuduPlannerTest uses another table.

Change-Id: I50664d29509716c8ac995fe8fbcee46233124c78
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6989
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: jenkins