For Kudu tables, delete the table's metadata if its not present in Kudu
Right now, if a given managed table is deleted in kudu through the client,
when a "DROP TABLE" statement is executed we fail with an exception and
the table remains in the meta store.

This leaves the user in a bad place since it can't recreate the table as a
managed table, as it is present in the meta store, but can't delete it
either since that operation fails.

This patch changes the behavior to issue a warning when the table is not
present in Kudu, but still delete the table from the metadata.

Change-Id: I17626dfb9dc7cbd0ee9a88382cffd4d2cdff861f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6990
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: jenkins