Backend part of Kudu predicate pushdown
This patch adds the optional ability for the backend KuduScanNode to
transform certain TExprs that it receives from the frontend into
KuduColumnRangePredicates.

The predicates are restricted to binary predicates of certain types
(only >=, <= and = are supported currently) which have a constant
on one side and a 'SlotRef' on the other.

This patch also adds several tests to KuduScanNodeTest that make sure
the predicates are indeed pushed down.

Change-Id: I964c7da2f9f93b591bf79e128d51b6f0b8f36dbf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7041
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins